extrapolate_image_from_excell/requirements.txt
openpyxl
openpyxl-image-loader
Pillow